Calculatefactorial

+++ Calculatefactorial.java
@@ -3,8 +3,8 @@ import java.util.Scanner;
 public class Calculatefactorial {
 
   public static void main(String[] args) {
-    int n=(new Scanner(System.in).nextInt());
-    int fact=1;
+    long n=(new Scanner(System.in).nextLong());
+    long fact=1;
     for(int i=1;i<=n;i++){
     fact=fact*i;
     }
